Create The Asylum's Curse
You are a private investigator trapped in an old, abandoned asylum. Rumor has it that a failed experiment turned one of the patients into a monstrous entity. Your goal is to find three keys scattered throughout the asylum to unlock the exit. The flashlight is your only source of light. If the monster finds you, it will start chasing you. Will you survive?